Safari-609.1.11.1/LayoutTests/editing/selection/ios/show-selection-in-transformed-container-expected.txt



Apply transform
This test verifies that after focusing a visible input field in a body element that has been translated horizontally out of view, the caret is still visible. To run the test manually, first tap the input field and check that the caret shows up. Then, tap the button to apply the CSS transform, tap the input field again, and check that the caret is still visible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS caretBefore.left is 20
PASS caretBefore.top is 14
PASS caretBefore.width is 2
PASS caretBefore.height is 23
PASS caretAfter.left is 20
PASS caretAfter.top is 14
PASS caretAfter.width is 2
PASS caretAfter.height is 23
PASS successfullyParsed is true

TEST COMPLETE
